feat : Design required components and made static code
With this change, All the required components are added. The components are also designed to have static code giving the idea of how they will look.
Next step is to add states and data flow to the code. The Component added are -
1) BookShelf
2) BookGrid
3) SearchBook

The component BookGrid is usable in both BookShelf and SearchBook.