feat: Added changing the shelf functionality
The functionality to change the shelf of the books is added. The function updateBook is added in App.js
This function updates the state books and also calls the API function 'update'.
The function 'updateBook' is passed as prop to following hierarchy - ListBooks -> BookShelf -> BookGrid.
From BookGrid the function called to change the state of component App. Inverse data flow is added.