feat : Get results according to search query and add them to shelf for later reading
Now the Compoenent SearchBook is complete. This component has two state, 'query' and 'searchBooks'.
query : It is the input string
searchBooks : this is an array of results to be deplayed.
When input text is updated. query is updated with setState. Later function 'updateBooksByQuery' is called which will update search results according to the new query.
'updateBook' in App component is changed to take care of case, when there is new book to be added to shelf, instead of just changing the shelf of existing book.