fix: Fixed the bug, which showed a book in 2 shelves after coming back from search page
The bug was - The book will be shown in two shelves, After the shelf of the book is changed from the search page.
This is now fixed by updating the function updateBook in App component. Now the book (for which shelf should be changed) is checked if present in the already existing books by matching IDs.