fix:  Fixed a bug in search page, where some search results were show after text is deleted
As described above, on the search page when all the search text is deleted, there were still some searches shown.
This is fixed by adding a lifecycle function componentDidUpadate. A condition is checked in this function, if query is empty but searchBooks is not empty then searchBooks is updated to be empty.